Adds Error comparisons + Tests
